mahekmirror - Slam - Copy/requirements.txt
aiohttp
anytree
aria2p
appdirs
attrdict
beautifulsoup4
cloudscrape
feedparser
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
js2py
lk21
lxml
pillow
psutil
psycopg2-binary
pybase64
pyrogram
pyshorteners
python-dotenv
python-magic
python-telegram-bot
qbittorrent-api
requests
selenium
speedtest-cli
telegraph
tenacity
TgCrypto
torrentool==1.1.0
urllib3
youtube_dl
ffmpeg
pytube
